requirements.txt
Pillow>=9.0.0
pystray>=0.18.0
